ReferLink_MVP_UI_Codebase/db/db_connect.py
import psycopg2
def get_connection():
    conn = psycopg2.connect(
        dbname="referlink",
        user="postgres",
        password="914Shift@",
        host="localhost",
        port="5432"
    )
    return conn
