ReferLink_MVP_UI_Codebase/requirements.txt
streamlit
streamlit-option-menu
fastapi
uvicorn
psycopg2-binary
pydantic
requests
